+++ skforecast/model_selection/tests/tests_search/test_evaluate_grid_hyperparameters.py
@@ -58,38 +58,6 @@ def test_TypeError_evaluate_grid_hyperparameters_when_cv_not_valid():
         )
 
 
-def test_TypeError_evaluate_grid_hyperparameters_when_forecaster_not_OneStepAhead():
-    """
-    Test TypeError is raised in _evaluate_grid_hyperparameters when forecaster is not
-    allowed to use OneStepAheadFold.
-    """
-
-    cv_one_step_ahead = OneStepAheadFold(
-            initial_train_size    = 100,
-            return_all_indexes    = False,
-        )
-    
-    class DummyForecaster:
-        pass
-    forecaster = DummyForecaster()
-    
-    err_msg = re.escape(
-        f"Only forecasters of type ['ForecasterRecursive', 'ForecasterDirect'] are allowed "
-        f"when using `cv` of type `OneStepAheadFold`. Got {type(forecaster).__name__}."
-    )
-    with pytest.raises(TypeError, match = err_msg):
-        _evaluate_grid_hyperparameters(
-            forecaster  = forecaster,
-            y           = y,
-            cv          = cv_one_step_ahead,
-            lags_grid   = [2, 4],
-            param_grid  = [{'alpha': 0.01}, {'alpha': 0.1}, {'alpha': 1}],
-            metric      = 'mean_absolute_error',
-            return_best = True,
-            verbose     = False
-        )
-
-
 def test_ValueError_evaluate_grid_hyperparameters_when_return_best_and_len_y_exog_different():
     """
     Test ValueError is raised in _evaluate_grid_hyperparameters when return_best