+++ docs/releases/releases.md
@@ -41,6 +41,8 @@ The main changes in this release are:
 
 + Added function `create_mean_pinball_loss` in the <code>[metrics]</code> module to create a function to calculate the mean pinball loss for a given quantile.
 
++ Added function `check_one_step_ahead_input` to check the input data when using a <code>[OneStepAheadFold]</code> in the <code>[model_selection]</code> functions.
+
 
 **Changed**